Gumbel distribution functions

Density, distribution function, quantile function and random generation for the gumbel distribtion. Extracted from polr source code

pgumbel(q, loc = 0, scale = 1, lower.tail = TRUE)

dgumbel(x, loc = 0, scale = 1, log = FALSE)

pGumbel(q, loc = 0, scale = 1, lower.tail = TRUE)

dGumbel(x, loc = 0, scale = 1, log = FALSE)

rgumbel(n, loc = 0, scale = 1)

rGumbel(n, loc = 0, scale = 1)

Arguments

q
vector of quantiles.

loc, scale
Location and scale parameters.

lower.tail
logical; if TRUE (default), probabilities are \(P[X \le x]\) otherwise, \(P[X > x]\).

x
vector of quantiles.

log
logical; if TRUE, probabilities p are given as log(p).

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Value

dgumbel gives the density, pgumbel gives the distribution function, and rgumbel generates random deviates.